test(events): group event unit tests
